Split Character selectors and maps
